pyproject.toml
[project]
name = "langgraph-tool-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11.9"
dependencies = [
    "chromadb>=1.0.15",
    "flask>=3.1.1",
    "langchain>=0.3.0",
    "langchain-chroma>=0.2.5",
    "langchain-community>=0.3.0",
    "langchain-openai>=0.3.28",
    "langgraph>=0.6.0",
    "langgraph-checkpoint>=2.1.1",
    "openai>=1.97.1",
    "pypdf>=5.9.0",
    "python-dotenv>=1.1.1",
    "tqdm>=4.67.1",
    "wikipedia>=1.4.0",
]

[tool.setuptools]
packages = ["vector_db_configuration"]
